+++ docker/image-auditor/Auditor/src/main/java/heig/dai/junodhaeffner/auditor/Auditor.java
@@ -1,0 +1,11 @@
+package heig.dai.junodhaeffner.auditor;
+
+import java.util.List;
+
+public class Auditor {
+
+    private static List<Musician> musicians;
+    public static void main(String[] args) {
+
+    }
+}
