+++ docker/image-auditor/Auditor/src/main/java/heig/dai/junodhaeffner/auditor/Musician.java
@@ -1,5 +1,6 @@
 package heig.dai.junodhaeffner.auditor;
 
 public class Musician {
-    f
+
+
 }
